requirements.txt
Django==5.2.5
djangorestframework==3.15.2
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
